+ another .file anme to convert at extraction time. * rewrote the code that creates the initial password to be more robust. * stupid typo in getHex() produced wrong hex reprs.

+++ arch/android/src/fr/inria/hop/hop.java
@@ -3,6 +3,7 @@ package fr.inria.hop;
 import java.io.BufferedOutputStream;
 import java.io.File;
 import java.io.FileOutputStream;
+import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.Enumeration;
@@ -11,6 +12,7 @@ import java.util.zip.ZipEntry;
 import java.util.zip.ZipFile;
 import java.math.BigInteger;
 import java.util.Date;
+import java.lang.String;
 
 import android.app.Activity;
 import android.os.Bundle;
@@ -72,9 +74,14 @@ public class hop extends Activity {
          while (entries.hasMoreElements()) {
             ZipEntry entry = (ZipEntry) entries.nextElement();
             String path = entry.getName().substring(zipFilterLength);
+            // convert dot.afile into .afile
             if (path.endsWith (dot_afile)) {
-               path = path.substring(0, path.length()-dot_afile.length())+".afile";
+               path = path.replace (dot_afile, ".afile");
                Log(".afile: "+path);
+            }
+            if (path.endsWith ("hoprc.hop")) {
+               path = path.replace ("config", ".config");
+               Log(".config: "+path);
             }
             File outputFile = new File(mAppRoot, path);
             File parent = outputFile.getParentFile();
@@ -149,37 +156,67 @@ public class hop extends Activity {
       return ans;
    }
 
+   public static void write (FileOutputStream fos, String line) throws IOException {
+      fos.write (line.getBytes ());
+      Log(line);
+   }
+
    public static String createAdminUser () {
-      // we create the hoprc.hop file with an initial admin user
+      // we create the wizard.hop file with an initial admin user
       // whose password is randomly generated
+      File wizard_hop= new File (mAppRoot+"/home/.config/hop/wizard.hop");
+      String password= null;
+      String scrambled= null;
 
-      Log("Creating admin user");
-      // 32bits==4bytes==8 hexa digits, 4Gi combinations
-      String password= new BigInteger (32, new java.util.Random ()).toString (16);
-      String scrambled= crypt ("admin", password);
-      Log("pass:  " + password);
-      Log("scram: " + scrambled);
+      if (! wizard_hop.exists ()) {
+         Log("Creating admin user");
+         // 32bits==4bytes==8 hexa digits, 4Gi combinations
+         password= new BigInteger (32, new java.util.Random ()).toString (16);
+         scrambled= crypt ("admin", password);
+         Log("pass:  " + password);
+         Log("scram: " + scrambled);
 
-      if (scrambled!=null) {
-         File wizard_hop= new File (mAppRoot+"/home/.config/hop/wizard.hop");
+         if (scrambled!=null) {
+            try {
+               FileOutputStream fos= new FileOutputStream (wizard_hop);
+               Date now= new Date ();
+               String date= now.toString ();
 
-         try {
-            FileOutputStream fos= new FileOutputStream (wizard_hop);
-            Date now= new Date ();
-            String date= now.toString ();
+               write (fos, ";; generated file, Hop installer "+date+"\n");
+               write (fos, ";; anonymous user\n");
+               write (fos, "(add-user! \"anonymous\")\n");
 
-            fos.write ((";; generated file, Hop installer "+date+"\n").getBytes ());
-            fos.write (";; anonymous user\n".getBytes ());
-            fos.write ("(add-user! \"anonymous\")\n".getBytes ());
-
-            fos.write (";; admin\n".getBytes ());
-            fos.write (("(add-user! \"admin\" :groups (quote (admin exec)) :password \""+scrambled+"\" :directories (quote *) :services (quote *))\n").getBytes ());
-         } catch (IOException e) {
-            // if we can't set the password
+               write (fos, ";; admin\n");
+               write (fos, "(add-user! \"admin\" :groups (quote (admin exec)) :password \""+scrambled+"\" :directories (quote *) :services (quote *))\n");
+            } catch (IOException e) {
+               Log ("can't write "+wizard_hop.getPath ()+": "+e);
+               // if we can't set the password
+               password= null;
+            }
+         } else {
+            Log("couldn't generate the scrambled password; maybe there's no md5 algrithm?");
             password= null;
          }
       } else {
-         password= null;
+         // we're screwed
+         try {
+            FileInputStream fis= new FileInputStream (wizard_hop);
+            byte data[] = new byte[BUFSIZE];
+            int count;
+
+            Log("here's the config");
+            while ((count = fis.read(data, 0, BUFSIZE)) != -1) {
+               String foo= "";
+               // Log(String (data));
+               for (int i=0; i<count; i++) {
+                  foo+= data[i];
+               }
+               Log(foo);
+            }
+         } catch (IOException e) {
+            // we can't even read the file
+            Log ("can't even read "+wizard_hop.getPath ()+": "+e);
+         }
       }
 
       // return the admin password
@@ -188,7 +225,7 @@ public class hop extends Activity {
 
    // copied from http://www.rgagnon.com/javadetails/java-0596.html
    // CC by-nc-sa Réal Gagnon <real@rgagnon.com>
-   static final String HEXES = "0123456789abcedf";
+   static final String HEXES = "0123456789abcdef";
    public static String getHex( byte [] raw ) {
       if ( raw == null ) {
          return null;
